The maximimum frame time is now clipped to 2G us to fix a bug where if you make the slice too long it would wrap around to a negative slilce value (which is an int)
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@3338 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/net/sf/jaer/graphics/AEPlayer.java
@@ -424,11 +424,11 @@ public class AEPlayer extends AbstractAEPlayer implements AEFileInputStreamInter
         aeFileInputStream.mark();
     }
 
-    public int position (){
+    public long position (){
         return aeFileInputStream.position();
     }
 
-    public void position (int event){
+    public void position (long event){
         aeFileInputStream.position(event);
     }
